svn: subclass Integer from long, some small fixes to Float

+++ sympy/sandbox/core/py_integer.py
@@ -3,13 +3,13 @@ from basic import Basic, sympify
 from number import Rational
 
 
-class Integer(Rational, int):
+class Integer(Rational, long):
 
     is_integer = True
 
     @memoizer_immutable_args('Integer.__new__')
     def __new__(cls, p):
-        obj = int.__new__(cls, p)
+        obj = long.__new__(cls, p)
         if p==0:
             obj.is_zero = True            
         elif p==1:
@@ -30,7 +30,7 @@ class Integer(Rational, int):
         if self is other: return 0
         c = cmp(self.__class__, other.__class__)
         if c: return c
-        return int.__cmp__(self, int(other))
+        return long.__cmp__(self, long(other))
 
     # converter methods
 
@@ -43,11 +43,11 @@ class Integer(Rational, int):
 
     @property
     def is_even(self):
-        return int.__mod__(self,2)==0
+        return long.__mod__(self,2)==0
 
     @property
     def is_odd(self):
-        return int.__mod__(self,2)==1
+        return long.__mod__(self,2)==1
 
     @property
     def is_positive(self):
